feat:[ASSMT-151]: :zap: Optimise Reconcile Check Code
