+++ pipelines.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"sort"
 	"strconv"
-	"strings"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
@@ -416,144 +415,4 @@ func createSpinnakerPipelines(pipelines interface{}) (reqId string, err error) {
 	reconcilePipeline(resp, queryParams)
 	log.Info("Spinnaker migration completed")
 	return reqId, nil
-}
-
-func reconcilePipeline(resp ResponseBody, queryParams map[string]string) {
-	var result map[string]interface{}
-	jsonData, err := json.Marshal(resp)
-	if err != nil {
-		log.Fatalf("Error occurred during marshalling. %v", err)
-	}
-	err = json.Unmarshal([]byte(jsonData), &result)
-	if err != nil {
-		log.Fatalf("Error occurred during unmarshalling. %v", err)
-	}
-	var pipelineID string
-	successfullyMigratedDetails := result["resource"].(map[string]interface{})["successfullyMigratedDetails"].([]interface{})
-	for _, detail := range successfullyMigratedDetails {
-		detailMap := detail.(map[string]interface{})
-		ngEntityDetail := detailMap["ngEntityDetail"].(map[string]interface{})
-		if ngEntityDetail["entityType"].(string) == "PIPELINE" {
-			pipelineID = ngEntityDetail["identifier"].(string)
-			break
-		}
-	}
-
-	if pipelineID == "" {
-		log.Fatalf("Pipeline ID not found in response")
-	}
-	//log.Info("Reconciliation started for pipeline with identifier : " + pipelineID)
-	uuid, err := getPipelineUUID(pipelineID, queryParams)
-	if err != nil {
-		log.Fatalf("Error getting pipeline UUID: %v", err)
-	}
-	//log.Info("Pipeline UUID : " + uuid)
-	reconcileNeeded, err := checkReconcileNeeded(uuid, queryParams)
-	if err != nil {
-		log.Fatalf("Error checking if reconcile is needed: %v", err)
-	}
-	if reconcileNeeded {
-		log.Info("Reconciliation is needed")
-		pipelineYaml, err := getPipelineYaml(pipelineID, queryParams)
-		if err != nil {
-			log.Fatalf("Error getting pipeline YAML: %v", err)
-		}
-		//log.Info("Pipeline YAML : " + pipelineYaml)
-		refreshedYaml, err := refreshPipelineYaml(pipelineYaml, queryParams)
-		if err != nil {
-			log.Fatalf("Error refreshing pipeline YAML: %v", err)
-		}
-		//log.Info("Refreshed Pipeline YAML : " + refreshedYaml)
-		success, err := updatePipelineYaml(pipelineID, refreshedYaml, queryParams)
-		if err != nil {
-			log.Fatalf("Error updating pipeline: %v", err)
-		}
-		if !success {
-			log.Fatalf("Failed to update pipeline")
-		}
-	} else {
-		log.Info("No reconciliation needed")
-	}
-}
-func getPipelineUUID(identifier string, queryParams map[string]string) (string, error) {
-	queryParams["getDefaultFromOtherRepo"] = "true"
-	url := GetUrlWithQueryParams(migrationReq.Environment, PipelineService, fmt.Sprintf("api/pipelines/%s/validate", identifier), queryParams)
-	respBody, err := Post(url, migrationReq.Auth, nil)
-	if err != nil {
-		return "", fmt.Errorf("failed to make request: %v", err)
-	}
-	uuid, ok := respBody.Data.(map[string]interface{})["uuid"].(string)
-	if !ok {
-		return "", errors.New("UUID not found in response")
-	}
-	return uuid, nil
-}
-
-func checkReconcileNeeded(uuid string, queryParams map[string]string) (bool, error) {
-	url := GetUrlWithQueryParams(migrationReq.Environment, PipelineService, fmt.Sprintf("api/pipelines/validate/%s", uuid), queryParams)
-	respBodyObj, err := Get(url, migrationReq.Auth)
-	if err != nil {
-		return false, fmt.Errorf("failed to make request: %v", err)
-	}
-	data, ok := respBodyObj.Data.(map[string]interface{})
-	if !ok {
-		return false, errors.New("data not found in response")
-	}
-	validateResp, ok := data["validateTemplateReconcileResponseDTO"].(map[string]interface{})
-	if !ok {
-		return false, errors.New("validateTemplateReconcileResponseDTO not found in response")
-	}
-	reconcileNeeded, ok := validateResp["reconcileNeeded"].(bool)
-	if !ok {
-		return false, errors.New("reconcileNeeded not found in response")
-	}
-	return reconcileNeeded, nil
-}
-
-func getPipelineYaml(identifier string, queryParams map[string]string) (string, error) {
-	queryParams["validateAsync"] = "true"
-	url := GetUrlWithQueryParams(migrationReq.Environment, PipelineService, fmt.Sprintf("api/pipelines/%s", identifier), queryParams)
-	respBodyObj, err := Get(url, migrationReq.Auth)
-	if err != nil {
-		return "", fmt.Errorf("failed to make request: %v", err)
-	}
-	data, ok := respBodyObj.Data.(map[string]interface{})
-	if !ok {
-		return "", errors.New("data not found in response")
-	}
-	yaml, ok := data["yamlPipeline"].(string)
-	if !ok {
-		return "", errors.New("yaml not found in response")
-	}
-	return yaml, nil
-}
-
-func refreshPipelineYaml(yaml string, queryParams map[string]string) (string, error) {
-	url := GetUrlWithQueryParams(migrationReq.Environment, TemplateService, "api/refresh-template/refreshed-yaml", queryParams)
-	respBodyObj, err := Post(url, migrationReq.Auth, map[string]string{"yaml": yaml})
-	if err != nil {
-		return "", fmt.Errorf("failed to make request: %v", err)
-	}
-	data, ok := respBodyObj.Data.(map[string]interface{})
-	if !ok {
-		return "", errors.New("data not found in response")
-	}
-	refreshedYaml, ok := data["refreshedYaml"].(string)
-	if !ok {
-		return "", errors.New("refreshedYaml not found in response")
-	}
-	return refreshedYaml, nil
-}
-
-func updatePipelineYaml(pipelineID string, yamlContent string, params map[string]string) (bool, error) {
-	url := GetUrlWithQueryParams(migrationReq.Environment, PipelineService, fmt.Sprintf("api/pipelines/v2/%s", pipelineID), params)
-	yamlReader := strings.NewReader(yamlContent)
-	respBodyObj, err := Put(url, migrationReq.Auth, yamlReader)
-	if err != nil {
-		return false, fmt.Errorf("error updating pipeline: %v", err)
-	}
-	if respBodyObj.Status == "SUCCESS" {
-		return true, nil
-	}
-	return false, fmt.Errorf("failed to update pipeline: %s", respBodyObj.Message)
 }
